+++ src/Ergosfare.Core/Extensions/TypeExtensions.cs
@@ -4,6 +4,7 @@ namespace Ergosfare.Core.Extensions;
 
 internal static class TypeExtensions
 {
+    
     public static IEnumerable<Type> GetInterfacesEqualTo(this Type type, Type interfaceType)
     {
         return type.GetInterfaces()
@@ -19,5 +20,5 @@ internal static class TypeExtensions
 
     public static IReadOnlyCollection<string> GetGroupsFromAttribute(this Type type)
         => ((GroupAttribute?)Attribute
-            .GetCustomAttribute(type, typeof(GroupAttribute)))?.GroupNames ?? [];
+            .GetCustomAttribute(type, typeof(GroupAttribute)))?.GroupNames ?? [GroupAttribute.DefaultGroupName];
 }